Lab 2 Initial Changes
I just did a lot of quick simple changes that immediately made sense to me with the lab including adding the title, page wrap and some of the sections. I need to fix the images, along with the formatting of my list. I also still need to get the nav bar in and working.