Try/Catch
Added some try, catch blocks so that my code doesn't break if something really weird happens with the coordinates as for some of the restaurants don't have coordinates like Mod pizza, so the map would permanently break when I tried to map to nonexistent coordinates. Still have to add more but this is a good start.